roll dancing strings examples into one
