+++ crates/egui_demo_lib/src/demo/dancing_strings.rs
@@ -1,9 +1,19 @@
-use egui::{containers::*, *};
+use egui::{containers::*, epaint::PathStroke, *};
+
+static GRADIENT: [Color32; 5] = [
+    hex_color!("#5BCEFA"),
+    hex_color!("#F5A9B8"),
+    Color32::WHITE,
+    hex_color!("#F5A9B8"),
+    hex_color!("#5BCEFA"),
+];
 
 #[derive(Default)]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 #[cfg_attr(feature = "serde", serde(default))]
-pub struct DancingStrings {}
+pub struct DancingStrings {
+    colors: bool,
+}
 
 impl super::Demo for DancingStrings {
     fn name(&self) -> &'static str {
@@ -27,6 +37,10 @@ impl super::View for DancingStrings {
         } else {
             Color32::from_black_alpha(240)
         };
+
+        ui.horizontal(|ui| ui.checkbox(&mut self.colors, "Show Colors"));
+
+        ui.separator();
 
         Frame::canvas(ui.style()).show(ui, |ui| {
             ui.ctx().request_repaint();
@@ -55,7 +69,37 @@ impl super::View for DancingStrings {
                     .collect();
 
                 let thickness = 10.0 / mode as f32;
-                shapes.push(epaint::Shape::line(points, Stroke::new(thickness, color)));
+                shapes.push(epaint::Shape::line(
+                    points,
+                    if self.colors {
+                        PathStroke::new_uv(thickness, move |_r, p| {
+                            let time = time / 10.0;
+                            let x = remap(p.x, rect.x_range(), 0.0..=1.0) as f64;
+                            let y = remap(p.y, rect.y_range(), 0.0..=1.0) as f64;
+
+                            let amp = (time * speed * mode).sin() / mode;
+                            let sin = amp * (time * std::f64::consts::TAU / 2.0 * mode).sin();
+
+                            let value = x * sin + y * sin;
+
+                            let color = if value < 0.2 {
+                                GRADIENT[0]
+                            } else if value < 0.4 {
+                                GRADIENT[1]
+                            } else if value < 0.6 {
+                                GRADIENT[2]
+                            } else if value < 0.8 {
+                                GRADIENT[3]
+                            } else {
+                                GRADIENT[4]
+                            };
+
+                            Color32::from_rgba_premultiplied(color[0], color[1], color[2], color[3])
+                        })
+                    } else {
+                        PathStroke::new(thickness, color)
+                    },
+                ));
             }
 
             ui.painter().extend(shapes);
