+++ crates/egui_demo_lib/src/demo/mod.rs
@@ -9,7 +9,6 @@ pub mod code_editor;
 pub mod code_example;
 pub mod context_menu;
 pub mod dancing_strings;
-pub mod dancing_trans_strings;
 pub mod demo_app_windows;
 pub mod drag_and_drop;
 pub mod extra_viewport;
